Do most of the conversion of the timer widget to hooks
You know, these hooks are starting to grow on me. useRef is ever-so-slightly more intuitive than instance ref properties (which weren't that intuitive before). 

Being able to write custom hook functions is also pretty cool, and has made the whole saving of state so much easier to understand.

Also sidenote: it seems like there's a bit of magic going on with IndexedDB, where it will allow multiple modioficattions to a row before it commits it. I know that it has a timed cache sort of thing, which actually seems to be being used here. It's really useful, as it means I don't need to implement something like this myself (which I was going to do)